[EGSVC-164] Workflow for new connection screen

+++ wcms/wcms-connection/src/main/java/org/egov/wcms/transanction/model/Connection.java
@@ -108,7 +108,7 @@ public class Connection {
     private String status;
 
     @NotNull
-    private int stateId;
+    private Long stateId;
 
     @NotNull
     private double sumpCapacity;
